Moved utilities to utils, made cfgwriter read a proper file, formatted function to snake_case (#4)
* Moved cfg-writer into manim/utils.
(It's a utility, so fits better there)

Edited setup.py accordingly.

* Moved config_utils into utils
(More of a utility, so fits better in utils)

* Edited logger.py to work with changes made.

Renamed parseTheme to parse_theme.

* Made config read defaults with absolute path.

* Use config_utils.successfully_read_files
